third_party/WebKit/LayoutTests/fast/dom/tabindex-behavior.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<input id="elementWithTabIndex" tabindex="3">
</body>
<script>
test(function(){
    var e = document.getElementById('elementWithTabIndex');
    assert_equals(e.tabIndex, 3);
    e.setAttribute("tabindex", "12039810238102983");
    assert_equals(e.getAttribute("tabindex"), "12039810238102983");
    assert_equals(e.tabIndex, 0);
    e.tabIndex = 3;
    e.setAttribute("tabindex", "-12039810238102983");
    assert_equals(e.getAttribute("tabindex"), "-12039810238102983");
    assert_equals(e.tabIndex, 0);
}, 'Track that setting tabIndex to out of range values the reflected value is 0.');
</script>
